Complete end-to-end pipeline: modality console, HL7 bridge, order poller hardening
Modality console web UI (new — sidecar/modality_console_web.py)
- Browser-based rad-tech worklist at port 5001, replacing the desktop GUI
- One-click "Image the Patient" patches source DICOM with MWL demographics
  and uploads to Orthanc modality; acquired state persists across tab switches
  via localStorage
- Sortable columns with Order Created timestamp; modality filter tabs (All/CR/US/CT)
- Flask added to sidecar Dockerfile; launched as 4th thread in main.py

HL7 bridge: MLLP → REST (sidecar/hl7_bridge.py)
- Switched from MLLP (port 8066 not active in PIH distro) to POST /ws/rest/v1/hl7
- XSS filter bypass: & in MSH-2 sent as \u0026 JSON Unicode escape
- Fixed MSH-4 (Sending Facility) to IMLADRIS — OpenMRS source lookup requires it

Order poller hardening (sidecar/order_poller.py)
- Modality map extended with PIH EMR parenthetical suffixes: (US), (CT), (MR)
- Order type discovery retries indefinitely (was 5 attempts) — startup order independent
- Removed autoExpireDate filter: PIH OpenMRS sets autoExpireDate = dateActivated,
  causing all new orders to appear expired immediately
- Query workaround: activatedOnOrAfterDate only (PIH rejects orderType without patient);
  filter to radiology orders client-side
- First-run lookback capped at 24 h to avoid replaying old history
- Follow-up patient fetch for DOB/sex when missing from order response
- Boundary timestamp advanced +1 s to prevent re-processing on next poll

Other sidecar fixes
- acquisition_loop.py: POLL_INTERVAL_MINUTES parsed as float (was int, crashed on 0.25)
- docker-compose.yml: demo polling intervals (15 s / 10 s), host.docker.internal for
  OpenMRS URL, hospital-records volume, modality console port 5001

Demo tooling
- demo/demo-instructions.md: full startup/teardown runbook with clean-state checklist
- tools/clear_demo_orders.py: void all pending radiology orders via OpenMRS REST API
- tools/clear_hl7_queue.py: purge HL7 inbound queue via MySQL (REST API is write-only)
- tools/regenerate_processed.py: rebuild botsabelo_processed DICOMs from current patient IDs
- tools/fulfill_order.py, restart_radiology_app.py, fix_provider_role.py: operational utils

+++ sidecar/acquisition_loop.py
@@ -20,7 +20,7 @@ logging.basicConfig(
 )
 log = logging.getLogger("acquisition_loop")
 
-POLL_MINUTES = int(os.getenv("POLL_INTERVAL_MINUTES", "5"))
+POLL_MINUTES = float(os.getenv("POLL_INTERVAL_MINUTES", "5"))
 
 
 def run_cycle():
